Clean up module naming and organization
