+++ boa_engine/src/vm/opcode/binary_ops/macro_defined.rs
