+++ boa_engine/src/vm/opcode/define/class/method.rs
@@ -1,0 +1,83 @@
+use crate::{
+    property::PropertyDescriptor,
+    vm::{opcode::Operation, ShouldExit},
+    Context, JsResult, JsString,
+};
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub(crate) struct DefineClassMethodByName;
+
+impl Operation for DefineClassMethodByName {
+    const NAME: &'static str = "DefineClassMethodByName";
+    const INSTRUCTION: &'static str = "INST - DefineClassMethodByName";
+
+    fn execute(context: &mut Context) -> JsResult<ShouldExit> {
+        let index = context.vm.read::<u32>();
+        let object = context.vm.pop();
+        let value = context.vm.pop();
+        let object = if let Some(object) = object.as_object() {
+            object.clone()
+        } else {
+            object.to_object(context)?
+        };
+        value
+            .as_object()
+            .expect("method must be function object")
+            .borrow_mut()
+            .as_function_mut()
+            .expect("method must be function object")
+            .set_home_object(object.clone());
+        let name = context.vm.frame().code.names[index as usize];
+        let name = context.interner().resolve_expect(name);
+        object.__define_own_property__(
+            name.into_common::<JsString>(false).into(),
+            PropertyDescriptor::builder()
+                .value(value)
+                .writable(true)
+                .enumerable(false)
+                .configurable(true)
+                .build(),
+            context,
+        )?;
+        Ok(ShouldExit::False)
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub(crate) struct DefineClassMethodByValue;
+
+impl Operation for DefineClassMethodByValue {
+    const NAME: &'static str = "DefineClassMethodByName";
+    const INSTRUCTION: &'static str = "INST - DefineClassMethodByName";
+
+    fn execute(context: &mut Context) -> JsResult<ShouldExit> {
+        let value = context.vm.pop();
+        let key = context.vm.pop();
+        let object = context.vm.pop();
+        let object = if let Some(object) = object.as_object() {
+            object.clone()
+        } else {
+            object.to_object(context)?
+        };
+        value
+            .as_object()
+            .expect("method must be function object")
+            .borrow_mut()
+            .as_function_mut()
+            .expect("method must be function object")
+            .set_home_object(object.clone());
+        let key = key.to_property_key(context)?;
+        object.__define_own_property__(
+            key,
+            PropertyDescriptor::builder()
+                .value(value)
+                .writable(true)
+                .enumerable(false)
+                .configurable(true)
+                .build(),
+            context,
+        )?;
+        Ok(ShouldExit::False)
+    }
+}
+
