+++ boa_engine/src/vm/opcode/define/class/mod.rs
@@ -1,0 +1,8 @@
+
+pub(crate) mod method;
+pub(crate) mod getter;
+pub(crate) mod setter;
+
+pub(crate) use method::*;
+pub(crate) use getter::*;
+pub(crate) use setter::*;
