+++ boa_engine/src/vm/opcode/mod.rs
@@ -5,13 +5,12 @@ use crate::{vm::ShouldExit, Context, JsResult};
 pub(crate) mod await_stm;
 pub(crate) mod call;
 pub(crate) mod concat;
-pub(crate) mod conversion;
+pub(crate) mod to;
 pub(crate) mod copy;
 pub(crate) mod define;
 pub(crate) mod delete;
 pub(crate) mod dup;
 pub(crate) mod environment;
-pub(crate) mod general;
 pub(crate) mod generator;
 pub(crate) mod get;
 pub(crate) mod iteration;
@@ -28,18 +27,22 @@ pub(crate) mod switch;
 pub(crate) mod throw;
 pub(crate) mod try_catch;
 pub(crate) mod value;
+pub(crate) mod jump;
+pub(crate) mod binary_ops;
+pub(crate) mod unary_ops;
+pub(crate) mod nop;
+pub(crate) mod void;
 
 // Operation structs
 pub(crate) use await_stm::*;
 pub(crate) use call::*;
 pub(crate) use concat::*;
-pub(crate) use conversion::*;
+pub(crate) use to::*;
 pub(crate) use copy::*;
 pub(crate) use define::*;
 pub(crate) use delete::*;
 pub(crate) use dup::*;
 pub(crate) use environment::*;
-pub(crate) use general::*;
 pub(crate) use generator::*;
 pub(crate) use get::*;
 pub(crate) use iteration::*;
@@ -56,6 +59,11 @@ pub(crate) use switch::*;
 pub(crate) use throw::*;
 pub(crate) use try_catch::*;
 pub(crate) use value::*;
+pub(crate) use jump::*;
+pub(crate) use binary_ops::*;
+pub(crate) use unary_ops::*;
+pub(crate) use nop::*;
+pub(crate) use void::*;
 
 pub(crate) trait Operation {
     const NAME: &'static str;
