+++ boa_engine/src/vm/opcode/nop/mod.rs
@@ -1,0 +1,16 @@
+use crate::{
+    vm::{opcode::Operation, ShouldExit},
+    Context, JsResult, JsValue,
+};
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub(crate) struct Nop;
+
+impl Operation for Nop {
+    const NAME: &'static str = "Nop";
+    const INSTRUCTION: &'static str = "INST - Nop";
+
+    fn execute(_context: &mut Context) -> JsResult<ShouldExit> {
+        Ok(ShouldExit::False)
+    }
+}
